% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdvancedAnalysis-class.R
\docType{class}
\name{AdvancedAnalysis-class}
\alias{AdvancedAnalysis-class}
\title{S4 Advanced Analysis Class}
\value{
An Advanced Analysis object
}
\description{
An Advanced Analysis object that is a parent class to all advanced analysis algorithms.
This class should never be called explicitly. Its purpose is to act as a parent class for advanced analyses.
}
\section{Slots}{

\describe{
\item{\code{channels}}{the channels selected for the advanced analysis, this can be either a list of short channel IDs (integer) OR long channel names (character)}

\item{\code{compensation_id}}{the compensation ID selected for the advanced analysis}

\item{\code{name}}{the name of the advanced analysis}

\item{\code{source_experiment}}{the source experiment ID the advanced analysis is associated with}

\item{\code{status}}{character representing the status of the advanced analysis}

\item{\code{.available_channels}}{the list of available channels based off the \link{panels.list} function}

\item{\code{.available_files}}{the list of available files based off the \link{fcs_files.list} function}

\item{\code{.available_populations}}{the list of available populations based off the \link{populations.list} function}
}}

