% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{news}
\alias{news}
\alias{CytobankAPI_news}
\title{News}
\usage{
CytobankAPI_news()
}
\description{
Get news on CytobankAPI updates
}
\details{
\code{CytobankAPI_news} View a log of CytobankAPI updates and release notes.
}
